pyproject.toml
[project]
name = "safar-sar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk==1.18.0",
]
